backend/requirements.txt
# TruthLens AI Backend - Python Dependencies

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.0

# Data Validation & Serialization
pydantic==2.0.3
pydantic-settings==2.0.3
python-json-logger==2.0.7

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Caching & Session Management
redis==5.0.0
aioredis==2.0.1

# Message Queue
pika==1.3.2
aiokafka==0.10.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-core==2.14.0

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
WordCloud==1.9.3

# Sentiment Analysis
vader-sentiment==3.3.2
transformers==4.35.0
torch==2.0.1

# Machine Learning & Data Science
scikit-learn==1.3.2
lightgbm==4.0.1
xgboost==2.0.0
numpy==1.24.3
pandas==2.1.2

# Audio Processing
librosa==0.10.0
pydub==0.25.1
soundfile==0.12.1

# Speech Recognition & LLM Integration
openai==1.3.0
google-cloud-speech==2.21.0
requests==2.31.0

# LLM Integration
openai==1.3.0
google-generativeai==0.3.0
anthropic==0.7.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
python-slugify==8.0.1

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0
prometheus-client==0.18.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
faker==20.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.13.0
mypy==1.7.0
pre-commit==3.5.0

# AWS Integration
boto3==1.29.5
botocore==1.32.5
aws-xray-sdk==2.12.0

# Performance & Async
aiofiles==23.2.1
asyncio-contextmanager==1.0.0
greenlet==3.0.1

# Time & Date
python-dateutil==2.8.2

# API Documentation
python-multipart==0.0.6
